Fix main.js
not sure if this is just me. maybe my school blocked these dns's, in which case feel free to deny this pull, but the dns stopped working for me so i replaced them with omada's dns. also, let's be honest, most people are using this in the us so i switched the order of the assigning of dns's